requirements.txt
Flask
google-generativeai
gTTS
